+++ test/typescript/apis/project/cds-test.ts
@@ -24,4 +24,12 @@ test.cds.ql === cds.ql
 test.cds.ApplicationService === cds.ApplicationService
 test.cds.get === cds.get
 
+cds.test.log().clear()
+cds.test.log().output.charAt(1)
+cds.test.log().release()
+
 const {server, url} = await test
+
+GET `/processor/$metadata`
+GET ('/processor/$metadata')
+GET ('/processor/$metadata', { auth: { username: 'alice', password: '' } })